pyproject.toml
[project]
name = "daily_agent"
version = "0.1.0"
description = "An agent to control the Daily platform"
authors = [
    {name = "Brandon Hancock", email = "brandon@brandonhancock.io"}
]
readme = "README.md"
requires-python = ">=3.12"
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = ">=3.12"
requests = "^2.31.0"
google-adk = "^0.1.0"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
google-cloud-aiplatform = {extras = ["adk", "agent_engines"], version = "^1.42.1"}
absl-py = "^2.1.0"
cloudpickle = "^3.0.0"
google-cloud-secret-manager = "^2.18.2"
deprecated = "^1.2.18"
psycopg2-binary = "^2.9.10"

[tool.poetry.scripts]
daily_agent = "daily_agent:app"
deploy-local = "deployment.local:main"
deploy-remote = "deployment.remote:main"
cleanup = "deployment.cleanup:cleanup_deployment"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

 